Created seperate branch for JavaFX
About time, too.
