+++ src/com/hydrogennx/JavaFXMainTest.java
@@ -1,0 +1,32 @@
+package com.hydrogennx;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+
+import java.io.IOException;
+import java.net.URL;
+
+public class JavaFXMainTest extends Application {
+
+    public static void main(String[] args) {
+        launch(args);
+    }
+
+    @Override
+    public void start(Stage primaryStage) throws IOException {
+
+        URL resource = getClass().getResource("JavaFXTest.fxml");
+        System.out.println(resource);
+        Parent mainWindow = FXMLLoader.load(resource);
+
+        primaryStage.setTitle("First Jump JavaFX MainMenu");
+        primaryStage.setScene(new Scene(mainWindow));
+        primaryStage.show();
+
+
+    }
+
+}
