+++ src/com/hydrogennx/javafx/JavaFXTest.java
@@ -1,7 +1,9 @@
 package com.hydrogennx.javafx;
 
 import javafx.application.Application;
+import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
+import javafx.scene.Node;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
@@ -9,25 +11,18 @@ import javafx.stage.Stage;
 import java.io.IOException;
 import java.net.URL;
 
-public class JavaFXTest extends Application {
+public class JavaFXTest {
 
-    public static void main(String[] args) {
-        launch(args);
-    }
-
-    @Override
-    public void start(Stage primaryStage) throws IOException {
-
-        System.out.println("Hello World!");
-
+    public static Node getMainWindow() throws IOException {
         URL resource = getClass().getResource("JavaFXTest.fxml");
         System.out.println(resource);
         Parent mainWindow = FXMLLoader.load(resource);
-
-        primaryStage.setTitle("First Jump JavaFX MainMenu");
-        primaryStage.setScene(new Scene(mainWindow));
-        primaryStage.show();
-
-
+        return mainWindow;
     }
+
+    @FXML
+    public void sayHi() {
+        System.out.println("Hello world!");
+    }
+
 }
